backend/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
Django==5.2.2
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.3.1
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
sqlparse==0.5.3
Werkzeug==3.1.3
